Set up folders for different interfaces, classes and services to more easily create a scalable appilication.
Set up the Car, CarFeatures and CarType class with a first set of data for object creation.

Set up a preliminary pricing service that indicates general structure a pricing service could have, although with probable changes to implementation.

Set up the User class as well as Interfaces for implementation later.